+++ domain/src/main/java/com/mument_android/domain/repository/mypage/BlockUserListRepository.kt
@@ -1,4 +1,4 @@
-package com.mument_android.domain.repository
+package com.mument_android.domain.repository.mypage
 
 import com.mument_android.domain.entity.mypage.BlockUserEntity
 import kotlinx.coroutines.flow.Flow
